Added basic structure, PracticeSessions and PracticeGroups
A practice session is a session with a date, a duration, and some notes
in a string format. A practice group is a name of the group and a
map of sessions.

As an example, a practice group might be "coding" or "violin", then all
practices of that group can be kept in one map and not repeat the name
of the practiced thing again and again for each object. I chose to store
these as a map with the key being dates for easy retrieval/removal of
specific dates.