Change library from simdjson to nlohmann for serialization
Attempting to serialize a string or a stream into a JSON object is not
what simdjson is built for. It is the fastest parser, since this project
requires more it will have to be a tradeoff in speed for writing JSON
objects to a file, which is 50% of the intended functionality.