Simple menu for showing group/session, making groups
main has been extended to include a menu for making groups, and showing
sessions and groups. The second the program launches it reads in all
groups for fast handling of them while the program is running. If the
use-case turns out to be many large groups, this is an unnecessary load
that will need to be remedied. As this is simply for testing purposes,
it will not matter for the final version.

It still needs to be possible to add a Session to a group.