Add session handling, refactoring
The main method has been refactored to instead have a pointer that gets
passed around when viewing a group and its details. It is from this
pointer now possible to list, add, and remove elements in the form of
practice sessions. This is done by adding code to make a practice
session using a static method that attempts to create one from an input
stream.

Save to file is now automatically ran for groups when they are finished
with an operation, such as on creation, or when adding/removing
sessions. Removing a session now throws an error if the date does not
exist. There also exists a method to get the name of the current group,
as that is stored in a private variable (as this is directly linked to
the name of the Json file).